refactor: move query artifacts into oprf_query.rs
