+++ crates/proof/README.md
@@ -13,7 +13,8 @@ Each ZK artifact type has exactly one way of being obtained:
 - **Circom artifacts** (zkeys, witness graphs) are trusted-setup outputs and cannot be
   rebuilt. `build.rs` uses files committed in this repository when present (development),
   and otherwise downloads them from the GitHub release tag pinned in `build.rs`.
-  Material loaders verify them against SHA-256 fingerprints pinned in `src/proof.rs`.
+  Material loaders verify them against SHA-256 fingerprints pinned in `src/oprf_query.rs`
+  (query proof) and `src/nullifier_proof.rs` (nullifier proof).
 - **Noir ownership proof artifacts** (`ownership_proof.pkp` / `.pkv`) are always built
   ad-hoc by `build.rs` from the checked-in circuit source using `nargo`. The required
   `nargo` version is pinned (see `flake.nix` and `REQUIRED_NARGO_VERSION` in `build.rs`); the build
@@ -65,7 +66,7 @@ The release tag is intentionally separate from the crate/software version track,
 - `circom/OPRFNullifier.arks.zkey`
 
 Material loaders verify artifacts against the SHA-256 fingerprints pinned in
-`src/proof.rs`.
+`src/oprf_query.rs` and `src/nullifier_proof.rs`.
 
 ### Publishing circuit artifact releases
 
@@ -73,7 +74,7 @@ Releases are created manually via the `Release circuit artifacts` GitHub Actions
 with a tag like `circuit-artifacts-v0.1.0`. The workflow creates a GitHub release and
 attaches the Circom artifact files listed above (committed in this repository). When
 publishing a new tag, update both the pinned tag in `build.rs` and the pinned SHA-256
-fingerprints in `src/proof.rs`.
+fingerprints in `src/oprf_query.rs` and `src/nullifier_proof.rs`.
 
 ## Noir ownership proof
