+++ crates/proof/src/artifacts/filesystem.rs
@@ -6,9 +6,8 @@ use std::{
 use crate::{
     OwnershipProver, OwnershipVerifier,
     artifacts::{ZkArtifactError, ZkArtifactKind, ZkArtifactSource},
-    nullifier_proof::{
-        CircomGroth16Material, load_nullifier_material_from_paths, load_query_material_from_paths,
-    },
+    nullifier_proof::{CircomGroth16Material, load_nullifier_material_from_paths},
+    oprf_query::load_query_material_from_paths,
 };
 
 use crate::ownership_proof::{load_ownership_prover_from_path, load_ownership_verifier_from_path};
